Rm cast to BigInteger to be cast to long

+++ src/main/kotlin/no/nav/cv/eures/samtykke/SamtykkeRepository.kt
@@ -86,7 +86,7 @@ private open class JpaSamtykkeRepository(
     @Transactional
     override fun hentAntallSamtykker() : Long
             = (entityManager.createNativeQuery(hentAntallSamtykker)
-            .singleResult as BigInteger).toLong()
+            .singleResult as Long)
 
 
     @Transactional
